+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 Unreleased
 ----------
+- Added example illustrating basic usage of `Readline` objects in a
+  terminal application
 - Adjusted program to use Rust Edition 2018
 - Removed `#![deny(warnings)]` attribute and demoted lints prone to
   future changes from `deny` to `warn`